requirements.txt
Django==1.10.5
django-import-export==0.5.1
